backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

pydantic
numpy
Pillow

# Model inference (Keras .h5)
tensorflow
h5py
fastapi
uvicorn[standard]
python-multipart
pydantic
tensorflow
numpy
pillow
